backend/auth_services/requirements.txt
fastapi
uvicorn
pydantic
boto3
python-jose
passlib